认证服务-用户登录

+++ fim_auth/auth_api/internal/config/config.go
@@ -12,6 +12,6 @@ type Config struct {
 	}
 	Auth struct {
 		AccessSecret string
-		AccessExpire int64
+		AccessExpire int
 	}
 }
